requirements.txt
discord.py>=2.3
httpx>=0.27
pydantic>=2.7
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
python-dateutil>=2.9
yfinance>=0.2
openai>=1.43
